added updated files with fix for issue 1
Custom icon pngs + the required json, html, and js files were added. These files include changes to get rid of a broken image link in the pop up the shows the pie of the month. I had failed to upload the code files when this repo was first created on GitHub, meaning that the bug that this code fixes was never available for download here.